fix(cli): PR #61 리뷰 R1·R2·NN1 — artifact get 무음 절단, 낡은 도움말, 50MB 사전 검사
R1 (차단): artifact get 이 16 MiB 초과 본문을 조용히 자르고 exit 0 을 냈다.
client.Do 의 io.ReadAll(io.LimitReader(body, 16<<20)) 이 한도에서 깨끗한 EOF 를
돌려주므로 오류가 나지 않았고, ArtifactGet 은 그 바이트를 --out 에 쓰고
saved_to·size_bytes 를 채워 성공을 보고했다. 잘린 diff 로 review approve 가 나가면
종료 조건 agent_approval 이 손상된 입력으로 충족된다(FR-2.2).

- DoStream + Client.OpenArtifactContent 추가: 아티팩트 본문은 c.Do 를 우회해
  스트리밍으로 받는다. 메모리에 담지 않는다.
- writeStream: 목적지 디렉터리의 임시 파일에 쓰고 본문이 다 온 뒤에만 rename 한다.
  Content-Length 가 있으면 기록 바이트와 대조하고, 불일치·전송 실패는 exit 5
  (download_truncated / download_failed) + 임시 파일 삭제. 부분 파일을 남기지 않는다.
- 다운로드는 Config.Timeout(전체 요청 기한) 대신 ctx 로 묶는다. 30초 기한은 큰
  아티팩트를 느린 링크에서 잘라 같은 결함을 되살린다.
- 버퍼링 응답(JSON·problem)의 한도는 MaxJSONResponse 로 명명하고 **자르는 대신
  거부**하도록 고쳤다(response_too_large). 조용한 절단 경로를 남기지 않는다.

R2: cmd/colab/p2.go 의 decision record 사용법이 v0.3 문자열이라 삭제된
--options·--chosen 을 광고했다. 브리프 [2]에 이 명령이 실리므로 에이전트가 읽고
시도한다. --summary/--rationale 로 고쳤고 artifact 사용법도 실제 플래그로 맞췄다.

NN1: 50 MB 상한은 이미 os.Stat 크기로 업로드 전에 검사하고 있었다(서버 413 의존
아님). 사용처가 추적되지 않았다는 지적에 회귀 테스트를 붙였다.

테스트 (모두 고치기 전 코드에서 실패하는 것을 확인했다):
- TestArtifactGetLargeBodyIsNotTruncated / ...StreamsRatherThanBuffers —
  17 MiB·64 MiB 본문. 고치기 전엔 size_bytes=16777216 으로 통과했다.
- TestArtifactGetShortTransferFailsAndRemovesFile — 선언보다 적게 보내는 서버.
  exit 5 이고 목적지 디렉터리에 임시 파일도 남지 않는다.
- TestUsageTextAdvertisesOnlyRealFlags — 삭제된 플래그를 도움말이 광고하면 실패.
- TestAdvertisedFlagsAllParse — 도움말이 광고하는 모든 --flag 가 실제로 파싱되는지.
  첫 판은 flag 패키지의 자체 목록을 읽어 무음 통과했고(단일 대시 표기도 놓쳤다)
  실제 usage 줄을 읽도록 다시 썼다. 없는 플래그를 usage 에 심어 무는 것을 확인했다.
- TestArtifactSubmitRefusesOversizeBeforeUploading — sparse 파일로 50 MB+1 과
  정확히 50 MB. 초과분은 요청 0건으로 exit 2.

리뷰어가 남겼다는 cli/internal/colab/zz_reviewer_probe_test.go 는 워킹트리·인덱스·
어느 ref 에도 없었다. 위 회귀 테스트를 새로 썼다.

+++ cli/internal/colab/actions_p2_test.go
@@ -382,7 +382,7 @@ func TestArtifactGetOutFile(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	if res.SavedTo != dest || res.SizeBytes == nil || *res.SizeBytes != len(clienttest.ArtifactBody) {
+	if res.SavedTo != dest || res.SizeBytes == nil || *res.SizeBytes != int64(len(clienttest.ArtifactBody)) {
 		t.Fatalf("res = %+v", res)
 	}
 	got, err := os.ReadFile(dest)
